DARNN on GPU
Initial version of DARNN on GPU. Closes #12.

The `Flux` implementation of `LSTMCell` cannot be used for training this
network on the GPU, because the `CUDA` dispatch throws an error when trying to
train on both hidden and cell state. `FALSTM` and the like are just
plain copies that avoid the specialised `CUDA` dispatch.